:ambulance:refactor: 유효성 검사 추가 및 유효성 검사 관련 함수 한 클래스로 병합
