+++ src/main/java/calculator/controller/MainController.java
@@ -3,7 +3,7 @@ package calculator.controller;
 import calculator.model.Calculator;
 import calculator.model.CustomDelimiter;
 import calculator.model.Splitter;
-import calculator.validation.OperandValidator;
+import calculator.validation.Validator;
 import calculator.view.InputView;
 import calculator.view.OutputView;
 
@@ -12,13 +12,13 @@ public class MainController {
     private OutputView outputView;
     private Calculator calculator;
     private CustomDelimiter customDelimiter;
-    private OperandValidator inputValidator;
+    private Validator inputValidator;
 
     public MainController(){
         this.inputView = new InputView();
         this.outputView = new OutputView();
         this.calculator = new Calculator();
-        this.inputValidator = new OperandValidator();
+        this.inputValidator = new Validator();
     }
 
     public void start(){
