+++ src/main/java/calculator/model/Calculator.java
@@ -1,12 +1,12 @@
 package calculator.model;
 
-import calculator.validation.OperandValidator;
+import calculator.validation.Validator;
 
 public class Calculator {
     public String plus(String[] splitedNumbers){
         int sumOfNumbers = 0;
         for (String token : splitedNumbers) {
-            int number = OperandValidator.parseNumber(token);
+            int number = Validator.parseNumber(token);
             sumOfNumbers += number;
         }
         return String.valueOf(sumOfNumbers);
